+++ sdk/templatespecs/arm-templatespecs/src/templateSpecsClient.ts
@@ -20,7 +20,7 @@ import { TemplateSpecsClientOptionalParams } from "./models";
 
 export class TemplateSpecsClient extends coreClient.ServiceClient {
   $host: string;
-  subscriptionId: string;
+  subscriptionId?: string;
   apiVersion: string;
 
   /**
@@ -33,12 +33,26 @@ export class TemplateSpecsClient extends coreClient.ServiceClient {
     credentials: coreAuth.TokenCredential,
     subscriptionId: string,
     options?: TemplateSpecsClientOptionalParams
+  );
+  constructor(
+    credentials: coreAuth.TokenCredential,
+    options?: TemplateSpecsClientOptionalParams
+  );
+  constructor(
+    credentials: coreAuth.TokenCredential,
+    subscriptionIdOrOptions?: TemplateSpecsClientOptionalParams | string,
+    options?: TemplateSpecsClientOptionalParams
   ) {
     if (credentials === undefined) {
       throw new Error("'credentials' cannot be null");
     }
-    if (subscriptionId === undefined) {
-      throw new Error("'subscriptionId' cannot be null");
+
+    let subscriptionId: string | undefined;
+
+    if (typeof subscriptionIdOrOptions === "string") {
+      subscriptionId = subscriptionIdOrOptions;
+    } else if (typeof subscriptionIdOrOptions === "object") {
+      options = subscriptionIdOrOptions;
     }
 
     // Initializing default values for options
@@ -50,7 +64,7 @@ export class TemplateSpecsClient extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-templatespecs/2.1.1`;
+    const packageDetails = `azsdk-js-arm-templatespecs/3.0.0`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -103,7 +117,7 @@ export class TemplateSpecsClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2021-05-01";
+    this.apiVersion = options.apiVersion || "2022-02-01";
     this.templateSpecs = new TemplateSpecsImpl(this);
     this.templateSpecVersions = new TemplateSpecVersionsImpl(this);
     this.addCustomApiVersionPolicy(options.apiVersion);