+++ sdk/templatespecs/arm-templatespecs/CHANGELOG.md
@@ -1,15 +1,33 @@
 # Release History
+    
+## 3.0.0 (2023-11-14)
+    
+**Features**
 
-## 2.1.1 (Unreleased)
+  - Added operation TemplateSpecs.getBuiltIn
+  - Added operation TemplateSpecs.listBuiltIns
+  - Added operation TemplateSpecVersions.getBuiltIn
+  - Added operation TemplateSpecVersions.listBuiltIns
+  - Added Interface TemplateSpecsGetBuiltInOptionalParams
+  - Added Interface TemplateSpecsListBuiltInsNextOptionalParams
+  - Added Interface TemplateSpecsListBuiltInsOptionalParams
+  - Added Interface TemplateSpecVersionsGetBuiltInOptionalParams
+  - Added Interface TemplateSpecVersionsListBuiltInsNextOptionalParams
+  - Added Interface TemplateSpecVersionsListBuiltInsOptionalParams
+  - Added Type Alias TemplateSpecsGetBuiltInResponse
+  - Added Type Alias TemplateSpecsListBuiltInsNextResponse
+  - Added Type Alias TemplateSpecsListBuiltInsResponse
+  - Added Type Alias TemplateSpecVersionsGetBuiltInResponse
+  - Added Type Alias TemplateSpecVersionsListBuiltInsNextResponse
+  - Added Type Alias TemplateSpecVersionsListBuiltInsResponse
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Class TemplateSpecsClient has a new signature
+  - Interface TemplateSpecsListByResourceGroupNextOptionalParams no longer has parameter expand
+  - Interface TemplateSpecsListBySubscriptionNextOptionalParams no longer has parameter expand
+    
+    
 ## 2.1.0 (2022-11-23)
     
 **Features**
@@ -27,4 +45,4 @@
 
 ## 2.0.0 (2021-12-16)
 
-- The package of @azure/arm-templatespecs is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart ).
+- The package of @azure/arm-templatespecs is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/js-track2-quickstart).